api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
pycryptodome==3.18.0
Werkzeug==2.3.7
setuptools==65.5.0